requirements.txt
python-dotenv==1.0.1
streamlit==1.44.1
pytest==8.3.5
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
langchain-community==0.3.7
langchain-pinecone==0.2.0
pinecone==6.0.2
curl_cffi==0.7.3
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==5.1.0
PyMuPDF==1.24.14
psycopg[binary]==3.2.3
psycopg2-binary==2.9.10
pgvector==0.3.6
redis==5.2.1
